backend/requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.27.0
djoser>=2.2.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
